refactor: remover fluxo de notificações das telas
